Split oversized test files to comply with 1000-line limit
Refactor subscription tests (1214 lines) into tests/ directory with
five focused modules: core, filter_take, debounce_throttle,
terminal_events, and subscription_ext.

Refactor status_bar tests (1016 lines) into tests/ directory with
four focused modules: style_item, state, component, and messages.
