+++ src/component/status_bar/tests/mod.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+mod component;
+mod messages;
+mod state;
+mod style_item;
